added helper clas
