+++ app/src/main/java/org/groundplatform/android/ui/datacollection/tasks/AbstractTaskFragment.kt
@@ -38,6 +38,7 @@ import org.groundplatform.android.model.submission.isNotNullOrEmpty
 import org.groundplatform.android.model.submission.isNullOrEmpty
 import org.groundplatform.android.model.task.Task
 import org.groundplatform.android.ui.common.AbstractFragment
+import org.groundplatform.android.ui.datacollection.DataCollectionUiState
 import org.groundplatform.android.ui.datacollection.DataCollectionViewModel
 import org.groundplatform.android.ui.datacollection.components.ButtonAction
 import org.groundplatform.android.ui.datacollection.components.LoiNameDialog
@@ -249,7 +250,7 @@ abstract class AbstractTaskFragment<T : AbstractTaskViewModel> : AbstractFragmen
     dataCollectionViewModel.loiNameDialogOpen.value = true
     renderComposableDialog {
       // The LOI NameDialog should call `handleLoiNameSet()` to continue to the next task.
-      ShowLoiNameDialog(dataCollectionViewModel.loiName.value ?: "") {
+      ShowLoiNameDialog((dataCollectionViewModel.uiState as DataCollectionUiState.Ready).loiName) {
         handleLoiNameSet(loiName = it)
       }
     }
